+++ src/index.ts
@@ -6,7 +6,7 @@ import * as webui from './webui'
 export const name = pluginName
 export const inject = {
     required: ['chatluna', 'chatluna_storage'],
-    optional: ['console', 'server']
+    optional: ['console']
 }
 
 export interface Config extends PluginConfig {}
@@ -21,9 +21,7 @@ export function apply(ctx: Context, config: Config) {
 }
 
 export * from './types'
-export * from './a2a'
 export * from './delegation'
 export * from './gateway'
 export * from './providers'
-export * from './agentd'
 export { AgentNexusService }